+++ src/containers/pages/AssigmentMapWrapper/helpers/utils.tsx
@@ -1,3 +1,4 @@
+import { Dictionary } from '@onaio/utils';
 import { Feature } from '@turf/turf';
 import { History } from 'history';
 import { EventData, Map } from 'mapbox-gl';
@@ -100,4 +101,66 @@ export const onJurisdictionClick = (
     }
   }
   return handleJurisdictionClick;
+};
+
+/**
+ * Map component popup handler
+ * @param map - mapbox map instance
+ * @param event - map event object
+ */
+
+export const buildMouseMoveHandler = (map: Map, event: EventData) => {
+  const popup = new mapboxgl.Popup({
+    closeButton: false,
+  });
+  // Added this part to remove duplicate popup containers showing on map when hovering
+  const nodes: any = document.getElementsByClassName('mapboxgl-popup');
+  let node: HTMLDivElement;
+  if (nodes && nodes.length) {
+    for (node of nodes) {
+      if (node.style) {
+        node.style.display = 'none';
+      }
+    }
+  }
+  let content: string = '';
+  let feature: Feature | any = {};
+  popup.remove();
+  // grab underlying features from map
+  const features: any = event.target.queryRenderedFeatures(event.point) as Dictionary[];
+  //
+  if (!features.length) {
+    return;
+  }
+
+  // Change the cursor style to pointer if features exist.
+  map.getCanvas().style.cursor = features.length ? 'pointer' : 'grab';
+
+  // loop through features, setting the name property to content var
+  for (feature of features) {
+    if (feature && feature.properties && feature.properties.name) {
+      content += `<div class="jurisdiction-name"><center>${feature.properties.name}</center></div>`;
+    }
+    /**
+     * Break out of loop once content is set
+     * This helps in a case where we could have 2 or more identical features
+     */
+    if (content) {
+      break;
+    }
+  }
+  /**
+   * Add popup to map after content is set
+   */
+  if (content.length) {
+    popup
+      .setLngLat(map.unproject(event.point))
+      .setHTML(content)
+      .addTo(map);
+  } else {
+    // clear existing popups if content is empty
+    popup.remove();
+    return;
+  }
+  return true;
 };
