+++ src/containers/pages/FocusInvestigation/map/active/tests/handlers.test.tsx
@@ -1,5 +1,5 @@
 import { Popup } from 'mapbox-gl';
-import { buildMouseMoveHandler, buildOnClickHandler } from '../helpers/utils';
+import { buildOnClickHandler } from '../helpers/utils';
 
 const popupHandlerMock = buildOnClickHandler('');
 
@@ -84,105 +84,5 @@ describe('activeFIMap/helpers/utils.popup', () => {
     popupHandlerMock(mockMap, event as any);
 
     expect(addToMock).toHaveBeenCalledWith({});
-  });
-  it('shows popup on hover', () => {
-    const event = {
-      lngLat: {
-        lat: 15.065355545319008,
-        lng: 101.1799767158821,
-      },
-      originalEvent: {},
-      point: {
-        x: 463.5,
-        y: 477.1875,
-      },
-      target: {
-        queryRenderedFeatures: () => {
-          return [
-            {
-              geometry: {
-                coordinates: [
-                  [
-                    [101.16072535514832, 15.119824869285075],
-                    [101.15796539300004, 15.052626968000027],
-                    [101.16026588800008, 15.052683043000059],
-                    [101.16249336800007, 15.05279176700003],
-                    [101.16457351200006, 15.05287588400006],
-                    [101.16817184600006, 15.053010455000049],
-                    [101.16992435000009, 15.05309115600005],
-                    [101.16997576800009, 15.052226568000037],
-                    [101.17015797000005, 15.050841579000064],
-                    [101.17258921900009, 15.050799469000026],
-                    [101.17424499900005, 15.050698872000055],
-                    [101.17613469500009, 15.050485078000065],
-                    [101.17699562400009, 15.050437232000036],
-                  ],
-                ],
-                type: 'Polygon',
-              },
-              id: '8fb28715-6c80-4e2c-980f-422798fe9f41',
-              properties: {
-                geographicLevel: 3,
-                name: 'Two Three Two Release Village',
-                parentId: '872cc59e-0bce-427a-bd1f-6ef674dba8e2',
-                status: 'Active',
-                version: 0,
-              },
-              type: 'Feature',
-            },
-          ];
-        },
-      },
-      type: 'fill',
-    };
-
-    const mockedPopup = Popup;
-    const addToMock = jest.fn();
-    mockedPopup.prototype.setLngLat = (e: any) => {
-      expect(e).toEqual({
-        lat: 15.065355545319008,
-        lng: 101.1799767158821,
-      });
-      e.setHTML = (f: string) => {
-        expect(f).toEqual(
-          '<div class="jurisdiction-name"><center>Two Three Two Release Village</center></div>'
-        );
-        e.addTo = addToMock;
-        return e;
-      };
-      return e;
-    };
-
-    (global as any).mapboxgl = {
-      Popup: mockedPopup,
-    };
-
-    const mockStyle = {
-      cursor: 'pointer',
-    };
-
-    const mockMap: any = {
-      // mock unproject function
-      getCanvas: jest.fn().mockImplementation(() => {
-        return {
-          style: {
-            ...mockStyle,
-          },
-        };
-      }),
-      unproject: () => {
-        return {
-          lat: 15.065355545319008,
-          lng: 101.1799767158821,
-        };
-      },
-    };
-
-    buildMouseMoveHandler(mockMap, event as any);
-
-    expect(addToMock.mock.calls).toEqual([
-      [{ getCanvas: expect.any(Function), unproject: expect.any(Function) }],
-    ]);
-    expect(addToMock).toBeCalledTimes(1);
   });
 });
